Added events for other dkg pallets and refactor
