+++ src/handlers/dkg/dkgMetaData/types.ts
@@ -1,4 +1,5 @@
 import type { Codec } from "@polkadot/types-codec/types"
+import { PalletDkgMetadataEvent } from "@polkadot/types/lookup"
 
 export enum DKGMethod {
   SIGNATURE_THRESHOLD = "setSignatureThreshold",
@@ -11,4 +12,14 @@ export interface SignatureThresholdArgs {
 
 export interface KeygenThresholdArgs {
   newThreshold: Codec
+}
+
+export type DKGMetaDataEvent = {
+  PublicKeySubmitted: PalletDkgMetadataEvent["asPublicKeySubmitted"]
+  NextPublicKeySubmitted: PalletDkgMetadataEvent["asNextPublicKeySubmitted"]
+  NextPublicKeySignatureSubmitted: PalletDkgMetadataEvent["asNextPublicKeySignatureSubmitted"]
+  PublicKeyChanged: PalletDkgMetadataEvent["asPublicKeyChanged"]
+  PublicKeySignatureChanged: PalletDkgMetadataEvent["asPublicKeySignatureChanged"]
+  MisbehaviourReportsSubmitted: PalletDkgMetadataEvent["asMisbehaviourReportsSubmitted"]
+  RefreshKeysFinished: PalletDkgMetadataEvent["asRefreshKeysFinished"]
 }
