+++ src/handlers/dkg/dkgProposalHandler/types.ts
@@ -1,0 +1,6 @@
+import { PalletDkgProposalHandlerEvent } from "@polkadot/types/lookup"
+
+export type DKGProposalHandlerEvent = {
+  InvalidProposalSignature: PalletDkgProposalHandlerEvent["asInvalidProposalSignature"]
+  ProposalSigned: PalletDkgProposalHandlerEvent["asProposalSigned"]
+}
