+++ src/handlers/dkg/dkgProposals/types.ts
@@ -1,0 +1,15 @@
+import { PalletDkgProposalsEvent } from "@polkadot/types/lookup"
+
+export type DKGProposalsEvent = {
+  ProposerThresholdChanged: PalletDkgProposalsEvent["asProposerThresholdChanged"]
+  ChainWhitelisted: PalletDkgProposalsEvent["asChainWhitelisted"]
+  ProposerAdded: PalletDkgProposalsEvent["asProposerAdded"]
+  ProposerRemoved: PalletDkgProposalsEvent["asProposerRemoved"]
+  VoteFor: PalletDkgProposalsEvent["asVoteFor"]
+  VoteAgainst: PalletDkgProposalsEvent["asVoteAgainst"]
+  ProposalApproved: PalletDkgProposalsEvent["asProposalApproved"]
+  ProposalRejected: PalletDkgProposalsEvent["asProposalRejected"]
+  ProposalSucceeded: PalletDkgProposalsEvent["asProposalSucceeded"]
+  ProposalFailed: PalletDkgProposalsEvent["asProposalFailed"]
+  AuthorityProposersReset: PalletDkgProposalsEvent["asAuthorityProposersReset"]
+}
