+++ src/handlers/dkg/type.ts
@@ -24,37 +24,6 @@ export enum DKGMetaDataSection {
   RefreshKeysFinished = "RefreshKeysFinished",
 }
 
-export type DKGMetaDataEvent = {
-  PublicKeySubmitted: PalletDkgMetadataEvent["asPublicKeySubmitted"]
-  NextPublicKeySubmitted: PalletDkgMetadataEvent["asNextPublicKeySubmitted"]
-  NextPublicKeySignatureSubmitted: PalletDkgMetadataEvent["asNextPublicKeySignatureSubmitted"]
-  PublicKeyChanged: PalletDkgMetadataEvent["asPublicKeyChanged"]
-  PublicKeySignatureChanged: PalletDkgMetadataEvent["asPublicKeySignatureChanged"]
-  MisbehaviourReportsSubmitted: PalletDkgMetadataEvent["asMisbehaviourReportsSubmitted"]
-  RefreshKeysFinished: PalletDkgMetadataEvent["asRefreshKeysFinished"]
-}
-
-export class EventDecoder<T extends Record<string, any>> {
-  private readonly _data: unknown
-  private readonly _key: unknown
-
-  constructor(public readonly event: SubstrateEvent) {
-    this._data = event.event.data.toHuman() as unknown
-    this._key = event.event.method.toString() as unknown
-  }
-
-  public is<Key extends keyof T>(key: Key): boolean {
-    return this._key !== key
-  }
-
-  public as<Key extends keyof T>(key: Key): T[Key] {
-    if (this._key !== key) {
-      throw new Error(`Event key ${this._key} does not match ${key as string}`)
-    }
-    return this._data as T[Key]
-  }
-}
-
 export enum DKGProposalsSection {
   // Vote threshold has changed (new_threshold)
   ProposerThresholdChanged = "ProposerThresholdChanged",
