+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./getInsideCalls"
 export * from "./signatureThreshold"
+export * from "./eventDecoder"
